requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf_yasg==1.21.8